+++ src/openassetio-core/tests/hostApi/ManagerTest.cpp
@@ -335,7 +335,9 @@ TEMPLATE_TEST_CASE(
     (BatchElementErrorMapping<openassetio::EntityResolutionErrorBatchElementException,
                               ErrorCode::kEntityResolutionError>),
     (BatchElementErrorMapping<openassetio::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>)){
+                              ErrorCode::kInvalidPreflightHint>),
+    (BatchElementErrorMapping<openassetio::InvalidTraitSetBatchElementException,
+                              ErrorCode::kInvalidTraitSet>)) {
   namespace hostApi = openassetio::hostApi;
   using trompeloeil::_;
 
@@ -644,7 +646,9 @@ TEMPLATE_TEST_CASE(
     (BatchElementErrorMapping<openassetio::EntityResolutionErrorBatchElementException,
                               ErrorCode::kEntityResolutionError>),
     (BatchElementErrorMapping<openassetio::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>)){
+                              ErrorCode::kInvalidPreflightHint>),
+    (BatchElementErrorMapping<openassetio::InvalidTraitSetBatchElementException,
+                              ErrorCode::kInvalidTraitSet>)) {
   namespace hostApi = openassetio::hostApi;
   using trompeloeil::_;
 
@@ -978,7 +982,9 @@ TEMPLATE_TEST_CASE(
     (BatchElementErrorMapping<openassetio::EntityResolutionErrorBatchElementException,
                               ErrorCode::kEntityResolutionError>),
     (BatchElementErrorMapping<openassetio::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>)){
+                              ErrorCode::kInvalidPreflightHint>),
+    (BatchElementErrorMapping<openassetio::InvalidTraitSetBatchElementException,
+                              ErrorCode::kInvalidTraitSet>)) {
   namespace hostApi = openassetio::hostApi;
   using trompeloeil::_;