+++ src/openassetio-python/tests/package/test_batchelementerror.py
@@ -32,6 +32,7 @@ from openassetio import (
     EntityAccessErrorBatchElementException,
     EntityResolutionErrorBatchElementException,
     InvalidPreflightHintBatchElementException,
+    InvalidTraitSetBatchElementException,
 )
 
 
@@ -43,6 +44,7 @@ class Test_BatchElementError_ErrorCode:
         assert int(BatchElementError.ErrorCode.kEntityAccessError) == 131
         assert int(BatchElementError.ErrorCode.kEntityResolutionError) == 132
         assert int(BatchElementError.ErrorCode.kInvalidPreflightHint) == 133
+        assert int(BatchElementError.ErrorCode.kInvalidTraitSet) == 134
 
 
 class Test_BatchElementError_inheritance:
@@ -125,6 +127,7 @@ class Test_BatchElementError_equality:
         EntityAccessErrorBatchElementException,
         EntityResolutionErrorBatchElementException,
         InvalidPreflightHintBatchElementException,
+        InvalidTraitSetBatchElementException,
     ],
 )
 # Note: these tests are not sufficient to assert that C++ exceptions